requirements.txt
# MEV Dashboard Requirements
# Core Django
Django==3.2.25
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-decouple==3.8
whitenoise==6.6.0
dj-database-url==2.1.0

# Channels for WebSockets
channels==3.0.5
channels-redis==4.2.0

# Database
psycopg2-binary==2.9.9

# Caching and message queue
redis==5.0.1
django-redis==5.4.0

# Async task processing
celery==5.3.4

# Machine Learning
scikit-learn==1.3.2
xgboost==2.0.3
tensorflow==2.15.0
pandas==2.1.4
numpy==1.24.3

# Data visualization (optional, for development)
matplotlib==3.8.2
seaborn==0.13.0

# Blockchain integration
web3==6.15.1
eth-abi==4.2.1

# Frontend
django-htmx==1.17.3

# HTTP requests
requests==2.31.0

# Development and testing
pytest==7.4.3
pytest-django==4.7.0
black==23.12.1
flake8==6.1.0
isort==5.13.2
django-debug-toolbar==4.2.0
django-extensions==3.2.3

# Environment management
python-dotenv==1.0.0

# Database
psycopg2-binary==2.9.9

# Caching and message queue
redis==5.0.1
django-redis==5.4.0

# Async task processing
celery==5.3.4

# Machine Learning
scikit-learn==1.3.2
xgboost==2.0.3
tensorflow==2.15.0
pandas==2.1.4
numpy==1.24.3

# Data visualization (optional, for development)
matplotlib==3.8.2
seaborn==0.13.0

# Blockchain integration
web3==6.15.1
eth-abi==4.2.1

# Frontend
django-htmx==1.17.3

# HTTP requests
requests==2.31.0

# Development and testing
pytest==7.4.3
pytest-django==4.7.0
black==23.12.1
flake8==6.1.0
isort==5.13.2
django-debug-toolbar==4.2.0
django-extensions==3.2.3

# Environment management
python-dotenv==1.0.0# Vector database support
pgvector==0.1.8
